Address listings component review feedback
