+++ client/src/app/(pages)/layout.tsx
@@ -1,8 +1,8 @@
 import './_globals/globals.scss';
 import React from 'react';
 import { Poppins } from 'next/font/google';
-import Navbar from "@/app/components/Navbar/Navbar";
-import Footer from '../components/Footer/Footer';
+import Navbar from "@/components/Navbar/Navbar";
+import Footer from "@/components/Footer/Footer";
 
 const poppins = Poppins({
   subsets: ['latin'],
