+++ client/src/mock/listings.ts
@@ -1,0 +1,23 @@
+export const mockListings = [
+    {
+      _id: "1",
+      title: "ESP32 Board",
+      price: 15,
+      imageUrl: "",
+      status: "active" as const,
+    },
+    {
+      _id: "2",
+      title: "Raspberry Pi",
+      price: 45,
+      imageUrl: "",
+      status: "draft" as const,
+    },
+    {
+      _id: "3",
+      title: "Arduino Kit",
+      price: 30,
+      imageUrl: "",
+      status: "completed" as const,
+    },
+  ];
